updated mapbox variant to v1.1.3

+++ src/backends/ncnn/ncnnlib.cc
@@ -211,10 +211,10 @@ namespace dd
                 probs.push_back(values[1]);
 
                 APIData ad_bbox;
-                ad_bbox.add("xmin",values[2] * inputc.width());
-                ad_bbox.add("ymax",values[3] * inputc.height());
-                ad_bbox.add("xmax",values[4] * inputc.width());
-                ad_bbox.add("ymin",values[5] * inputc.height());
+                ad_bbox.add("xmin",static_cast<double>(values[2] * inputc.width()));
+                ad_bbox.add("ymax",static_cast<double>(values[3] * inputc.height()));
+                ad_bbox.add("xmax",static_cast<double>(values[4] * inputc.width()));
+                ad_bbox.add("ymin",static_cast<double>(values[5] * inputc.height()));
                 bboxes.push_back(ad_bbox);
             }
         }
